refactor: split funGames.ts into 6 domain files + barrel (#475)
Moves birds, bugsInsects, superheroes, artCraft, camping, fairyTaleChars
into funGamesData/ subfiles. funGames.ts becomes a 6-line barrel — zero
breaking changes for consumers.
